优化报错信息。

+++ src/OpensslAesServiceProvider.php
@@ -19,7 +19,7 @@ class OpensslAesServiceProvider extends ServiceProvider
             //秘钥
             $s_key = env('AES_KEY');
             $s_method = env('AES_METHOD', 'AES-256-CBC');
-            $i_offset = env('AES_OFFSET', 6);
+            $i_offset = (int) env('AES_OFFSET', 6);
 
             return new LumenOpensslAES($s_key, $s_method, $i_offset);
         });
